+++ bungee/src/main/java/com/github/games647/fastlogin/bungee/McAPIProfile.java
@@ -1,0 +1,15 @@
+package com.github.games647.fastlogin.bungee;
+
+public class McAPIProfile {
+
+    private String uuid;
+    private String name;
+
+    public String getUuid() {
+        return uuid;
+    }
+
+    public String getName() {
+        return name;
+    }
+}